Implemented 18-bot ODT addresses and related changes.

+++ src/kd/kdf11_a/odt/atprompt.cpp
@@ -39,7 +39,7 @@ KDF11_A_ODT::State KDF11_A_ODT::StateMachine::transition (AtPrompt_1 && currentS
         return move (currentState);
     }
 
-    if (context_->location_.isA<AddressLocation<u16>> ())
+    if (context_->location_.isA<AddressLocation<u32>> ())
     {
         // Print last opened address and it's value
         return context_->writeAddressContents (context_->location_.wordAddress ());
